--- src/app/auth/page.tsx
@@ -1,9 +1,0 @@
-import Register from "@/components/auth/Register";
-
-export default function AuthPage() {
-  return (
-    <main className="w-full min-h-screen flex items-center justify-center">
-      <Register />
-    </main>
-  );
-}
