+++ src/components/auth/AuthFooter.tsx
@@ -1,0 +1,12 @@
+export default function AuthFooter() {
+  return (
+    <>
+      <footer className="text-center py-6 border border-outline rounded-b-3xl bg-sub shadow-xl">
+        <div className="space-x-2">
+          <span>이미 계정이 있으신가요?</span>
+          <span className="text-text-4 underline">로그인</span>
+        </div>
+      </footer>
+    </>
+  );
+}
