--- src/components/auth/Login.tsx
@@ -1,7 +1,0 @@
-export default function Login() {
-  return (
-    <>
-      <h1>Login</h1>
-    </>
-  );
-}
