+++ src/components/landingPage/section/SixthSection.tsx
@@ -1,9 +1,12 @@
+"use client";
+
 import Button from "../../common/Button";
+import Link from "next/link";
 
 export default function SixthSection() {
   return (
     <>
-      <section className="relative w-full overflow-hidden">
+      <section id="start" className="relative w-full overflow-hidden py-60">
         <div className="max-w-80 mx-auto space-y-25">
           <div className="w-full text-center space-y-5">
             <h3 className="text-4xl font-bold leading-12">
@@ -17,10 +20,16 @@ export default function SixthSection() {
               특별한 순간을 만들어줄 것입니다
             </p>
             <div className="space-x-3">
-              <Button className="text-white">편지 쓰러 가기</Button>
-              <Button className="rounded-xl border-2 bg-white border-[#FF2600]/30">
-                로그인
-              </Button>
+              <Link href={"/capsules/new"}>
+                <Button className="bg-primary text-white">
+                  편지 쓰러 가기
+                </Button>
+              </Link>
+              <Link href={"/auth/login"}>
+                <Button className="rounded-xl border-2 bg-white border-primary/30">
+                  로그인
+                </Button>
+              </Link>
             </div>
           </div>
           <div className="text-[#525252] space-y-4 text-center">
