+++ src/components/landingPage/section/third/ThirdSection.tsx
@@ -3,7 +3,7 @@ import CardList from "./CardList";
 export default function ThirdSection() {
   return (
     <>
-      <section>
+      <section id="what" className="py-60">
         <CardList />
       </section>
     </>
