+++ src/type/auth.d.ts
@@ -1,0 +1,1 @@
+type AuthMode = "login" | "register";
